+++ quisp/rules/actions/SwappingAction.cc
@@ -8,7 +8,7 @@ namespace quisp {
 namespace rules {
 namespace actions {
 
-SwappingAction::SwappingAction(unsigned long RuleSet_id, int rule_index, int lp, QNIC_type lqt, int lqi, int lqad, int lr, int rp, QNIC_type rqt, int rqi, int rqad, int rr,
+SwappingAction::SwappingAction(unsigned long RuleSet_id, unsigned long rule_index, int lp, QNIC_type lqt, int lqi, int lqad, int lr, int rp, QNIC_type rqt, int rqi, int rqad, int rr,
                                int slqi, QNIC_type slqt, int srqi, QNIC_type srqt) {
   ruleset_id = RuleSet_id;
   rule_id = rule_index;
